Adapt to changes from tests-split in kivakit
c.f. https://github.com/Telenav/kivakit/pull/100
