+++ kivakit-filesystems/github/pom.xml
@@ -35,6 +35,12 @@
             <artifactId>github-api</artifactId>
         </dependency>
 
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>kivakit-test</artifactId>
+            <scope>test</scope>
+        </dependency>
+
     </dependencies>
 
 </project>
