+++ kivakit-filesystems/java/pom.xml
@@ -29,6 +29,12 @@
             <artifactId>kivakit-component</artifactId>
         </dependency>
 
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>kivakit-test</artifactId>
+            <scope>test</scope>
+        </dependency>
+
     </dependencies>
 
 </project>
