+++ kivakit-filesystems/s3fs/pom.xml
@@ -24,7 +24,7 @@
         <!-- KivaKit -->
 
         <dependency>
-            <groupId>com.telenav.kivakit</groupId>
+            <groupId>${project.groupId}</groupId>
             <artifactId>kivakit-settings</artifactId>
         </dependency>
 
@@ -35,6 +35,16 @@
             <artifactId>s3</artifactId>
         </dependency>
 
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>kivakit-test</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.telenav.kivakit</groupId>
+            <artifactId>kivakit-test-internal</artifactId>
+            <scope>test</scope>
+            <type>jar</type>
+        </dependency>
     </dependencies>
 
 </project>
