+++ kivakit-web/jetty/src/main/java/module-info.java
@@ -9,6 +9,9 @@ open module kivakit.web.jetty
     requires org.eclipse.jetty.util;
     requires org.eclipse.jetty.webapp;
     requires transitive javax.servlet.api;
+    requires transitive kivakit.test;
+    requires transitive kivakit.test.internal;
+    requires kivakit.network.http;
 
     // Module exports
     exports com.telenav.kivakit.web.jetty;
